refactor: DEFAULT AVATAR IMAGE를 contants로 분리
