+++ client/src/shared/constants/index.ts
@@ -1,1 +1,4 @@
 export const ACCESS_TOKEN_KEY = 'access_token';
+
+export const DEFAULT_AVATAR_URL =
+  'https://ahmadda-dev.s3.ap-northeast-2.amazonaws.com/profile_avatar.png';
